Rename main to misc to avoid confusion
